Keep lifecycle reconciliation behind domain boundary
