+++ api/openapi.yaml
@@ -5566,6 +5566,39 @@ paths:
             text/markdown:
               schema:
                 type: string
+  /grc/findings/{findingID}/security-lifecycle/reconcile:
+    post:
+      operationId: reconcileSecurityLifecycleFinding
+      summary: Reconcile one lifecycle finding against current source evidence
+      tags:
+        - GRC
+      parameters:
+        - $ref: '#/components/parameters/findingID'
+      requestBody:
+        required: true
+        content:
+          application/json:
+            schema:
+              type: object
+              required:
+                - tenant_id
+              properties:
+                tenant_id:
+                  type: string
+                  description: Authorized tenant that owns the lifecycle finding.
+      responses:
+        '200':
+          description: Durable finding state after an open observation or verified closure.
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/SecurityLifecycleFindingReconcileResponse'
+        '202':
+          description: Verification is pending because current provenance or source coverage is incomplete.
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/SecurityLifecycleFindingReconcileResponse'
   /grc/findings/{findingID}/audit-preview:
     get:
       operationId: previewGRCAuditPacket
@@ -9712,6 +9745,25 @@ components:
           items:
             type: object
             additionalProperties: true
+    SecurityLifecycleFindingReconcileResponse:
+      type: object
+      required: [finding_id, status, verification, audit_preview_url]
+      properties:
+        finding_id:
+          type: string
+          description: Stable lifecycle finding URN.
+        status:
+          type: string
+          enum: [open, resolved]
+        verification:
+          type: string
+          description: Current reconciliation outcome.
+        reason:
+          type: string
+          description: Operator action or evidence condition when verification is pending.
+        audit_preview_url:
+          type: string
+          description: Live audit-preview path for this exact durable finding.
     GRCFindingsResponse:
       type: object
       required: [findings, meta, generated_at]
